Restructures components
