+++ src/game/components/IndividualComponents/CarriedItemComponent.ts
@@ -1,0 +1,12 @@
+import { ComponentType } from '../Components';
+
+export type CarriedItemComponentProps = { item: string };
+
+export class CarriedItemComponent {
+  type = ComponentType.CarriedItem;
+  item: string;
+
+  constructor({ item }: CarriedItemComponentProps) {
+    this.item = item;
+  }
+}
