+++ src/game/components/IndividualComponents/HandlingComponent.ts
@@ -1,0 +1,5 @@
+import { ComponentType } from '../Components';
+
+export class HandlingComponent {
+  type = ComponentType.Handling;
+}
