+++ src/game/components/IndividualComponents/InteractingComponent.ts
@@ -1,0 +1,5 @@
+import { ComponentType } from '../Components';
+
+export class InteractingComponent {
+  type = ComponentType.Interacting;
+}
