+++ src/game/components/IndividualComponents/MovableComponent.ts
@@ -1,0 +1,5 @@
+import { ComponentType } from '../Components';
+
+export class MovableComponent {
+  type = ComponentType.Movable;
+}
