+++ src/game/components/IndividualComponents/PickableComponent.ts
@@ -1,0 +1,5 @@
+import { ComponentType } from '../Components';
+
+export class PickableComponent {
+  type = ComponentType.Pickable;
+}
