+++ src/game/components/IndividualComponents/PlayerComponent.ts
@@ -1,0 +1,5 @@
+import { ComponentType } from '../Components';
+
+export class PlayerComponent {
+  type = ComponentType.Player;
+}
