+++ src/game/components/IndividualComponents/VelocityComponent.ts
@@ -1,0 +1,14 @@
+import { ComponentType } from '../Components';
+
+export type VelocityComponentProps = { vx: number; vy: number };
+
+export class VelocityComponent {
+  type = ComponentType.Velocity;
+  vx: number;
+  vy: number;
+
+  constructor({ vx, vy }: VelocityComponentProps) {
+    this.vx = vx;
+    this.vy = vy;
+  }
+}
