+++ src/game/components/IndividualComponents/WalkableComponent.ts
@@ -1,0 +1,5 @@
+import { ComponentType } from '../Components';
+
+export class WalkableComponent {
+  type = ComponentType.Walkable;
+}
